Refactor default form success message detection to form factory
[MAILPOET-2810]
