+++ tests/integration/Form/FormFactoryTest.php
@@ -4,6 +4,7 @@ namespace MailPoet\Form;
 
 use MailPoet\Entities\FormEntity;
 use MailPoet\Form\Templates\TemplateRepository;
+use MailPoet\Settings\SettingsController;
 
 class FormFactoryTest extends \MailPoetTest {
 
@@ -37,6 +38,26 @@ class FormFactoryTest extends \MailPoetTest {
     expect($formEntity->getBody())->notEmpty();
     expect($formEntity->getSettings())->notEmpty();
     expect($formEntity->getStyles())->string();
+  }
+
+  public function testItSetsDefaultMessage() {
+    $settings = $this->diContainer->get(SettingsController::class);
+    $settings->set('signup_confirmation.enabled', true);
+    $formEntity = $this->formFactory->createFormFromTemplate(TemplateRepository::INITIAL_FORM_TEMPLATE);
+    $formSettings = $formEntity->getSettings() ?? [];
+    expect($formSettings['success_message'])->equals('Check your inbox or spam folder to confirm your subscription.');
+
+    $settings->set('signup_confirmation.enabled', false);
+    $formEntity = $this->formFactory->createFormFromTemplate(TemplateRepository::INITIAL_FORM_TEMPLATE);
+    $formSettings = $formEntity->getSettings() ?? [];
+    expect($formSettings['success_message'])->equals('You’ve been successfully subscribed to our newsletter!');
+
+    $formEntity = $this->formFactory->createFormFromTemplate(
+      TemplateRepository::INITIAL_FORM_TEMPLATE,
+      ['success_message' => 'My custom']
+    );
+    $formSettings = $formEntity->getSettings() ?? [];
+    expect($formSettings['success_message'])->equals('My custom');
   }
 
   public function testItCanOverrideTemplateSettings() {
